Extract process killing from Custom Executor
The custom executor has the ideal situation where we want to first send
SIGTERM to the process to allow the process to clean up. If the process
is still alive after a specific deadline it will send a SIGKILL as
explained in
https://docs.gitlab.com/runner/executors/custom.html#terminating-and-killing-executables

Extract this logic to it's package so it can be reused for executors
like the `shell` executor to solve issues like
https://gitlab.com/gitlab-org/gitlab-runner/issues/3376

+++ helpers/process/killer_unix.go
@@ -14,7 +14,7 @@ type unixKiller struct {
 	process *os.Process
 }
 
-func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
+func newKiller(logger common.BuildLogger, process *os.Process) killer {
 	return &unixKiller{
 		logger:  logger,
 		process: process,
@@ -22,6 +22,10 @@ func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
 }
 
 func (pk *unixKiller) Terminate() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Signal(syscall.SIGTERM)
 	if err != nil {
 		pk.logger.Errorln("Failed to send SIGTERM signal:", err)
@@ -32,6 +36,10 @@ func (pk *unixKiller) Terminate() {
 }
 
 func (pk *unixKiller) ForceKill() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Kill()
 	if err != nil {
 		pk.logger.Errorln("Failed to force-kill:", err)
