+++ helpers/process/killer_windows.go
@@ -11,7 +11,7 @@ type windowsKiller struct {
 	process *os.Process
 }
 
-func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
+func newKiller(logger common.BuildLogger, process *os.Process) killer {
 	return &windowsKiller{
 		logger:  logger,
 		process: process,
@@ -19,15 +19,24 @@ func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
 }
 
 func (pk *windowsKiller) Terminate() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Kill()
 	if err != nil {
 		pk.logger.Errorln("Failed to terminate:", err)
 
+		// try to kill right-after
 		pk.ForceKill()
 	}
 }
 
 func (pk *windowsKiller) ForceKill() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Kill()
 	if err != nil {
 		pk.logger.Errorln("Failed to force-kill:", err)
